Add `CompanyPolicy` on destroy method, and initiate it in `AuthServiceProvider` for Company model
- Add Company Policy with policy on destroy Company
- Load Company Policy in Auth Service Provider
- Use authorize in controller to check is logged user is authorized to destroy company

+++ app/Http/Controllers/CompanyController.php
@@ -103,12 +103,10 @@ class CompanyController extends Controller
      */
     public function destroy(Company $company)
     {
-        
-        /**
-         * Delete An Existing Company
-         */
+        $this->authorize('destroy', $company);
 
         $company->delete();
+        
         return redirect('/companies');
     }
